[readme] add versioning entry for current npm packages
identify a NODE version that 'works' with the npm packages that the repo
currently usages

I've confirmed that node v17.0.1 works such that the npm scripts run
with success.  I've tried using v17.9 --- but it did not work.
    'work' here means that 'npm install' was successful.
    'not work' here means that 'npm install' was unsuccessful.

it's plausible that some v17.{2-8} might work, but I've not tried
those... for now just need a solid version that works and allows me to
get going with repo changes.